Made components file and moved filed inside components file, ald installed prop types library
