+++ src/components/AddTodoForm.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import PropTypes from "prop-types";
 import InputWithLabel from "./InputWithLabel";
 
 function AddTodoForm({ onAddTodo }) {
@@ -52,5 +53,9 @@ function AddTodoForm({ onAddTodo }) {
     </div>
   );
 }
+
+AddTodoForm.propTypes = {
+  onAddTodo: PropTypes.func.isRequired,
+};
 
 export default AddTodoForm;
