+++ src/components/TodoListItem.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import PropTypes from "prop-types"; // Import PropTypes
 import style from "./TodoListItem.module.css";
 
 function TodoListItem({ title, id, onRemoveTodo }) {
@@ -9,5 +10,11 @@ function TodoListItem({ title, id, onRemoveTodo }) {
     </li>
   );
 }
+
+TodoListItem.propTypes = {
+  title: PropTypes.string.isRequired,
+  id: PropTypes.number.isRequired,
+  onRemoveTodo: PropTypes.func.isRequired, // Define propTypes
+};
 
 export default TodoListItem;
