+++ src/components/TodoListItem.module.css
@@ -1,7 +1,7 @@
 body {
   background-color: #caedab;
   color: #398908;
-  font-family: "Roboto", sans-serif; /* Added font family */
+  font-family: "Roboto", sans-serif;
 }
 .ListItem {
   color: #348190;
